:truck: :: Data 컴포넌트 분리
Index, Search, User 컴포넌트 1폴더 ->  각각 분리
